face-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
deepface==0.0.93
opencv-python-headless==4.10.0.84
Pillow==10.4.0
qdrant-client==1.11.3
google-api-python-client==2.147.0
google-auth==2.35.0
numpy==1.26.4
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
httpx==0.27.2
apscheduler==3.10.4
tf-keras
